# swagger.yaml
openapi: 3.0.0
paths:
...
  /inventory/stock:
    post:
      operationId: InventoriesController_upsertStock
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertStockDto"
      responses:
        "201":
          description: Збережений запис складу
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryStockDto"
      summary: Створити/оновити залишок для виду інвентарю
      tags: &a3
        - inventory
    get:
      operationId: InventoriesController_listStock
      parameters: []
      responses:
        "200":
          description: Список складу з доступністю
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSummaryDto"
      summary: Отримати total/available по всіх видах
      tags: *a3
  /inventory/issue:
    post:
      operationId: InventoriesController_issue
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueItemDto"
      responses:
        "201":
          description: Створено/оновлено позицію студента
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentInventoryDto"
      summary: Видати студенту певну кількість інвентарю
      tags: *a3
  /inventory/return:
    post:
      operationId: InventoriesController_return
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnItemDto"
      responses:
        "200":
          description: "Оновлена позиція студента або { closed: true }"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StudentInventoryDto"
                  - $ref: "#/components/schemas/CloseFlagDto"
      summary: Прийняти повне/часткове повернення
      tags: *a3
  /inventory/students/{studentId}/items:
    get:
      operationId: InventoriesController_listStudentItems
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            example: "1"
            type: number
      responses:
        "200":
          description: Масив активних позицій студента
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentInventoryDto"
      summary: Список активних позицій конкретного студента
      tags: *a3
...
  /inventories/audits/{id}:
    get:
      operationId: InventoryAuditsController_get
      parameters:
        - name: id
          required: true
          in: path
          description: UUID аудиту
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Деталі аудиту
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryAuditDto"
      summary: Отримати аудит за ID
      tags: *a4
    delete:
      operationId: InventoryAuditsController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: UUID аудиту
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Аудит успішно видалено
        "404":
          description: Аудит не знайдено
      summary: Видалити аудит за ID
      tags: *a4
info:
  title: Hostel API
  description: Документація інвентаризації
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    InventoryKind:
      type: string
      enum:
        - tulle
        - curtains
        - blanket
        - mattress
        - pillowcase
        - mattressCover
        - duvetCover
        - waffleTowel
        - terryTowel
        - sheet
        - cover
        - pillow
        - tablecloth
        - bedSet
    UpsertStockDto:
      type: object
      properties:
        kind:
          example: mattress
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        total:
          type: number
          example: 50
          minimum: 0
      required:
        - kind
        - total
    InventoryStockDto:
      type: object
      properties:
        kind:
          description: Вид інвентарю
          example: blanket
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        total:
          type: number
          description: Загальна кількість одиниць на складі
          example: 50
          minimum: 0
      required:
        - kind
        - total
    StockSummaryDto:
      type: object
      properties:
        kind:
          description: Вид інвентарю
          example: blanket
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        total:
          type: number
          description: Загальна кількість одиниць на складі
          example: 50
          minimum: 0
      required:
        - kind
        - total
    IssueItemDto:
      type: object
      properties:
        studentId:
          type: number
          example: 1
          minimum: 1
        kind:
          example: blanket
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        quantity:
          type: number
          example: 2
          minimum: 1
      required:
        - studentId
        - kind
        - quantity
    StudentBriefDto:
      type: object
      properties:
        id:
          type: number
          description: ID студента
          example: 2
        fullName:
          type: string
          description: ПІБ
          example: Іваненко Іван Іванович
        roomNumber:
          type: string
          description: Кімната
          example: A-21
        faculty:
          type: string
          description: Факультет
          example: ФІОТ
        course:
          type: number
          description: Курс
          example: 1
        studyGroup:
          type: string
          description: Навчальна група
          example: КП-12
      required:
        - id
        - fullName
        - roomNumber
        - faculty
        - course
        - studyGroup
    StudentInventoryDto:
      type: object
      properties:
        id:
          type: string
          description: UUID позиції
          example: ff8fd4f2-a5a5-4c69-b20b-4a4a5617a7c7
          format: uuid
        kind:
          description: Вид інвентарю
          example: towel
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        quantity:
          type: number
          description: Поточна кількість, видана студенту (залишок по позиції)
          example: 4
          minimum: 0
        issuedAt:
          type: string
          description: Час створення/першої видачі
          example: 2025-09-12T21:42:37.186Z
          format: date-time
        returnedAt:
          type: object
          description: Час повернення (null, якщо ще не повернено)
          example: null
          nullable: true
          format: date-time
        student:
          description: Дані студента-власника позиції
          allOf:
            - $ref: "#/components/schemas/StudentBriefDto"
      required:
        - id
        - kind
        - quantity
        - issuedAt
        - student
    CloseFlagDto:
      type: object
      properties:
        closed:
          type: boolean
          description: Ознака повного закриття позиції після повернення
          example: true
      required:
        - closed
    ReturnItemDto:
      type: object
      properties:
        studentId:
          type: number
          example: 1
          minimum: 1
        kind:
          example: blanket
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        quantity:
          type: number
          example: 1
          minimum: 1
          description: "Опціонально: якщо пропущено — повне повернення"
      required:
        - studentId
        - kind
    InventoryAuditItemDto:
      type: object
      properties:
        id:
          type: string
          description: UUID позиції аудиту
          example: 8bf4a0b9-2d24-4d1c-9a2e-1f7b4f6e4a10
        kind:
          description: Вид інвентарю
          example: blanket
          allOf:
            - $ref: "#/components/schemas/InventoryKind"
        total:
          type: number
          description: Загальна кількість на момент аудиту
          example: 120
          minimum: 0
        issued:
          type: number
          description: Кількість, видана студентам
          example: 37
          minimum: 0
        available:
          type: number
          description: Доступна кількість
          example: 83
          minimum: 0
      required:
        - id
        - kind
        - total
        - issued
        - available
    InventoryAuditDto:
      type: object
      properties:
        id:
          type: string
          description: UUID аудиту
          example: f115c75d-41ba-4657-bbff-7faff9404fce
        createdAt:
          type: string
          description: Мітка часу створення аудиту
          example: 2025-09-12T21:47:10.123Z
          format: date-time
        items:
          description: Знімок складу на момент аудиту
          example:
            - id: 1a2b3c4d-0000-0000-0000-000000000001
              kind: blanket
              total: 120
              issued: 37
              available: 83
            - id: 1a2b3c4d-0000-0000-0000-000000000002
              kind: mattress
              total: 60
              issued: 12
              available: 48
          type: array
          items:
            $ref: "#/components/schemas/InventoryAuditItemDto"
      required:
        - id
        - createdAt
        - items